Better handling of platforms when downloading
Check whether platform is supported before moving on

Add some tests to ensure it's the case. Also that we support
all said platform (file can actually be downloaded). This test suite
is slow, but I think it's worth it

Note: in future releases, windows support will need to be reworked.
in some instances, the `exe` is gone and replaced with a zip that'll
likely need to unpack