Improve the debugging experience
Add a `lspJavaOpt` setting to forward java-opt to coursier. This allows
different thing, but in this context it was to a add a debug options.

I've also added the `m2Local` coursier repository because coursier
does not support it by default anymore and gradle publishes there.
So if you use gradle, rather than mill, in the smithy-language-server
then you won't find locally published jar.

It was my mistake, I should never have removed it in the first place.